Capture more data

+++ Services/FileClient.cs
@@ -3,6 +3,7 @@ using Microsoft.Extensions.Options;
 
 using SixLabors.ImageSharp;
 using SixLabors.ImageSharp.Processing;
+using SixLabors.ImageSharp.Metadata.Profiles.Exif;
 
 using ImageProcessor.Models;
 
@@ -29,14 +30,14 @@ public class FileClient
             // Only return files with the extension specified in appsettings. 
             // Exclude any files with "_thumb." in their name to prevent processing thumbnails
             var fi = new FileInfo(x);
-            return _config.ImageExtensions.Contains(fi.Extension) && !fi.Name.Contains("_thumb.");
+            return _config.ImageExtensions.Contains(fi.Extension.ToLower()) && !fi.Name.Contains("_thumb.") && !fi.Name.Contains("_preview.");
         });
     }
 
     /// <summary>Returns an enumerable of subfolders that contain images</summary>
     public IEnumerable<string> GetSubDirectories(string directory)
     {
-        return Directory.EnumerateDirectories(directory).Where(x => this.GetImages(x).Any());
+        return Directory.EnumerateDirectories(directory);
     }
 
     /// <summary>
@@ -62,7 +63,7 @@ public class FileClient
             {
                 Category = tags.ElementAtOrDefault(0) ?? "",
                 SubCategory = tags.ElementAtOrDefault(1) ?? "",
-                Tags = (tags.ElementAtOrDefault(2) ?? "").Split(",").Select(x => x.Trim()).ToArray(),
+                Tags = (tags.ElementAtOrDefault(2) ?? "").Split(",").Select(x => x.Trim()).Where(x => !string.IsNullOrWhiteSpace(x)).ToArray(),
             };
         }
         else
@@ -79,18 +80,27 @@ public class FileClient
         }
     }
 
-    /// <summary>Loads the <see cref="file"/> using ImageSharp, generates a thumbnail image</summary>
+    /// <summary>Loads the <see cref="file"/> using ImageSharp, generates a thumbnail image and a preview image to reduce total filesize going over the wire</summary>
     public async Task<ImageData> CreateThumbnail(string file, TagData tags)
     {
         // Load the image
         var fi = new FileInfo(file);
         var img = await Image.LoadAsync(file);
+        var id = Guid.NewGuid().ToString();
 
-        // Name the thumbnail
-        string thumbnail = fi.FullName.Substring(0, fi.FullName.Length - fi.Extension.Length) + "_thumb" + fi.Extension;
+        // Name the resized images
+        string thumbnail = Path.Combine(fi.Directory?.FullName ?? "", id + "_thumb" + fi.Extension);
+        string preview = Path.Combine(fi.Directory?.FullName ?? "", id + "_preview" + fi.Extension);
 
         // Resize the image to be a max of 250 width and height, retaining the aspect ratio
-        float ratio = (img.Width > img.Height ? img.Width : img.Height) / _config.ThumbnailMaxSize;
+        float ratio = (img.Width > img.Height ? img.Width : img.Height) / _config.PreviewMaxSize;
+        img.Mutate(x => x.Resize((int)(img.Width / ratio), (int)(img.Height / ratio)));
+
+        // Save the preview
+        await img.SaveAsync(preview);
+
+        // Process thumbnail 
+        ratio = (img.Width > img.Height ? img.Width : img.Height) / _config.ThumbnailMaxSize;
         img.Mutate(x => x.Resize((int)(img.Width / ratio), (int)(img.Height / ratio)));
 
         // Save the thumbnail
@@ -98,10 +108,22 @@ public class FileClient
 
         return new ImageData()
         {
+            GUID = Guid.NewGuid().ToString(),
             Image = file,
             Thumbnail = thumbnail,
-            CreatedDate = fi.CreationTime,
-            MetaData = tags
+            Preview = preview,
+            Extension = fi.Extension,
+            CreatedDate = DateTime.ParseExact(img.Metadata.ExifProfile.GetValue<string>(ExifTag.DateTime).ToString() ?? "", "yyyy:MM:dd HH:mm:ss", null),
+            TagData = tags,
+            Camera = new CameraSettings()
+            {
+                Model = img.Metadata.ExifProfile.GetValue<string>(ExifTag.Model).Value,
+                Flash = img.Metadata.ExifProfile.GetValue<ushort>(ExifTag.Flash).Value,
+                ISO = img.Metadata.ExifProfile.GetValue<uint>(ExifTag.RecommendedExposureIndex).Value,
+                ShutterSpeed = SimplifyRational(img.Metadata.ExifProfile.GetValue<SixLabors.ImageSharp.Rational>(ExifTag.ExposureTime).Value),
+                Aperature = SimplifyRational(img.Metadata.ExifProfile.GetValue<SixLabors.ImageSharp.Rational>(ExifTag.FNumber).Value),
+                FocalLength = SimplifyRational(img.Metadata.ExifProfile.GetValue<SixLabors.ImageSharp.Rational>(ExifTag.FocalLength).Value)
+            }
         };
     }
 
@@ -115,7 +137,7 @@ public class FileClient
         string archiveDir = Directory.GetParent(input[0].destination.Image)?.FullName ?? "";
 
         // First validate all files are valid, so we don't move half of them and fail
-        if (input.All(x => File.Exists(x.source.Image) && File.Exists(x.source.Thumbnail)))
+        if (input.All(x => File.Exists(x.source.Image) && File.Exists(x.source.Thumbnail) && File.Exists(x.source.Preview)))
         {
             // Create a subfolder if needed.
             if (!Directory.Exists(archiveDir))
@@ -129,16 +151,29 @@ public class FileClient
                 _logger.LogInformation($"Moving {src.Image} to {dest.Image}");
                 File.Move(src.Image, dest.Image, _config.OverwriteFilesInArchive);
 
+                _logger.LogInformation($"Moving {src.Preview} to {dest.Preview}");
+                File.Move(src.Preview, dest.Preview, _config.OverwriteFilesInArchive);
+
                 _logger.LogInformation($"Moving {src.Thumbnail} to {dest.Thumbnail}");
                 File.Move(src.Thumbnail, dest.Thumbnail, _config.OverwriteFilesInArchive);
             }
 
             // And now delete the /tags file once the copy has been successful
-            if (_config.DeleteAutoTagFile)
+            if (File.Exists(Path.Combine(sourceDir, _config.AutoTagFileName)))
             {
-                File.Delete(
-                    Path.Combine(sourceDir, _config.AutoTagFileName)
-                );
+                if (_config.DeleteAutoTagFile)
+                {
+                    File.Delete(
+                        Path.Combine(sourceDir, _config.AutoTagFileName)
+                    );
+                }
+                else
+                {
+                    File.Move(
+                        Path.Combine(sourceDir, _config.AutoTagFileName),
+                         Path.Combine(archiveDir, _config.AutoTagFileName + DateTime.Now.ToString("yyyyMMdd_HHmmss"))
+                    );
+                }
             }
         }
         else
@@ -151,6 +186,33 @@ public class FileClient
                                     );
 
             throw new FileNotFoundException($"The following files could not be found: \r\n{string.Join("\r\n", missingFiles)}");
+        }
+    }
+
+    /// <summary>Deletes the specified folder if it's empty</summary>
+    public void DeleteDirectory(string directory)
+    {
+        // Delete the folder if it's empty. This will throw an exception if the directory isn't actually empty, but that shouldn't happen
+        if (directory != _config.SourceFolder && !Directory.GetFileSystemEntries(directory).Any())
+        {
+            _logger.LogInformation($"Deleting source folder {directory}");
+            Directory.Delete(directory);
+        }
+    }
+
+    /// <summary>
+    /// Formats a ImageSharp.Rational into a more viewable string. Changes "3000/10" to "300" and "10/2000" to "1/200"
+    /// </summary>
+    public string SimplifyRational(Rational input)
+    {
+        if (input.Denominator > input.Numerator)
+        {
+            return $"1/{input.Denominator / input.Numerator}";
+        }
+        else
+        {
+            return $"{(float)input.Numerator / (float)input.Denominator}";
+
         }
     }
 }
